Load medical history editor on demand
