+++ tests/playwright/cold-load-budget.spec.js
@@ -104,6 +104,9 @@ test('cold mobile app load stays within committed resource budgets', async ({ pa
     new URL(entry.name).pathname === '/js/context-card-lifestyle-editors-impl.js'
   ))).toBe(false);
   expect(entries.some(entry => (
+    new URL(entry.name).pathname === '/js/context-card-medical-history-editor-impl.js'
+  ))).toBe(false);
+  expect(entries.some(entry => (
     new URL(entry.name).pathname === '/js/sun-defaults.js'
   ))).toBe(false);
   expect(entries.some(entry => (
