+++ tests/test-correctness-phase2.js
@@ -139,6 +139,7 @@ const pwaAppShellAssets = [
   '/js/context-card-summaries.js',
   '/js/context-card-editor-ui.js',
   '/js/context-card-medical-history-editor.js',
+  '/js/context-card-medical-history-editor-impl.js',
   '/js/context-card-lifestyle-editors.js',
   '/js/context-card-lifestyle-editors-impl.js',
   '/js/supplement-impact.js',
