+++ tsconfig.checkjs.json
@@ -102,6 +102,7 @@
     "js/context-card-lifestyle-editors.js",
     "js/context-card-lifestyle-editors-impl.js",
     "js/context-card-medical-history-editor.js",
+    "js/context-card-medical-history-editor-impl.js",
     "js/context-card-summaries.js",
     "js/context-source-registry.js",
     "js/context-cards-runtime.js",
