+++ tsconfig.json
@@ -89,6 +89,7 @@
     "js/context-card-lifestyle-editors.js",
     "js/context-card-lifestyle-editors-impl.js",
     "js/context-card-medical-history-editor.js",
+    "js/context-card-medical-history-editor-impl.js",
     "js/context-card-summaries.js",
     "js/context-cards.js",
     "js/crypto.js",
